# docs/swagger.yaml
basePath: /api/v1
definitions:
  wallet.Broker:
    properties:
      CNPJ:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
  wallet.CertificateOfDeposit:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      dueDate:
        type: string
      fixedInterestRate:
        type: number
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - dueDate
    - fixedInterestRate
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
  wallet.FICFI:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
...
host: localhost:8889
info:
  contact:
    name: API Support
    url: https://github.com/mfinancecombr/finance-wallet-api
  description: mfinance Wallet API data.
  license:
    name: BSD 3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
  title: MFinance Wallet API
  version: 0.1.0
paths:
...
  /brokers/{id}:
    delete:
      consumes:
      - application/json
      description: delete some broker by id
      parameters:
      - description: Broker id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Delete broker by ID
    get:
      consumes:
      - application/json
      description: get all broker data
      parameters:
      - description: Broker id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Broker'
      summary: Get a broker
    put:
      consumes:
      - application/json
      description: Update some broker by id
      parameters:
      - description: Broker id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update broker data by ID
...
  /certificate-of-deposit/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get certificate of deposi purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.CertificateOfDeposit'
      summary: Get get certificate of deposit purchase by ID
    put:
      consumes:
      - application/json
      description: update new certificate of deposit purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some certificate of deposit purchase
...
  /certificates-of-deposit/sales/{id}:
    get:
      consumes:
      - application/json
      description: get certificate of deposit sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.CertificateOfDeposit'
      summary: Get get certificate of deposit sale by ID
    put:
      consumes:
      - application/json
      description: update new certificate of deposit sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some certificate of deposit sale
...
  /ficfi/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get FIFCI purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.FICFI'
      summary: Get FICFI purchase by ID
    put:
      consumes:
      - application/json
      description: update new FICFI purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some FICFI purchase
...
  /ficfi/sales/{id}:
    get:
      consumes:
      - application/json
      description: get FICFI sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.FICFI'
      summary: Get FICFI sale by ID
    put:
      consumes:
      - application/json
      description: update new FICFI sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some FICFI sale
...
swagger: "2.0"